// code-app/src/generated/models/OGBOriginationSharePointTransportModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

export interface ManualTriggerInput {
  // Requested transport operation: ensureFolder, upload, verifyFolder, or verifyFile
  text: string;
  // Please enter your input
  text_1: string;
  // Please enter your input
  text_2: string;
  // Please enter your input
  text_3: string;
  // Please enter your input
  text_4: string;
  // Please enter your input
  text_5: string;
  // Please enter your input
  text_6: string;
  // Please enter your input
  text_7: string;
  // Please enter your input
  text_8: string;
  // Please enter your input
  text_9: string;
  // Please enter your input
  text_10: string;
  // Please enter a number
  number: number;
  // Please select file or image
  file: {
    name?: string;
    contentBytes?: string;
  };
  // Canonical SHA-256 of the immutable governed request material computed by the Code App.
  text_11: string;
}

export interface ResponseActionOutput {
  transportresponse?: string;
}